added regenerating feature
